speech/requirements.txt
Flask==0.12.2
google-cloud-speech==0.29.0
google-cloud-storage==1.4.0
gunicorn==19.7.1